backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
alembic>=1.12.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
pytesseract>=0.3.10
pillow>=10.0.0
pdfplumber>=0.10.0
python-magic>=0.4.27
pdf2image>=1.16.3